requirements.txt
picamera==1.13
pkg-resources==0.0.0
RPi.GPIO==0.7.0
